+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/TestTask.java
@@ -21,6 +21,9 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import gov.healthit.chpl.dto.CertificationResultTestTaskDTO;
 import gov.healthit.chpl.dto.TestParticipantDTO;
 import gov.healthit.chpl.dto.TestTaskDTO;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Singular;
 
 /**
  * A task used for SED testing for a given criteria.
@@ -28,6 +31,8 @@ import gov.healthit.chpl.dto.TestTaskDTO;
 @XmlType(namespace = "http://chpl.healthit.gov/listings")
 @XmlAccessorType(XmlAccessType.FIELD)
 @JsonIgnoreProperties(ignoreUnknown = true)
+@Builder
+@AllArgsConstructor
 public class TestTask implements Serializable {
     private static final long serialVersionUID = -3761135258451736516L;
 
@@ -163,6 +168,7 @@ public class TestTask implements Serializable {
     /**
      * The set of criteria within a listing to which this task is applied.
      */
+    @Singular("criterion")
     @XmlElementWrapper(name = "criteriaList", nillable = true, required = false)
     @XmlElement(name = "criteria")
     private Set<CertificationCriterion> criteria;
@@ -172,6 +178,7 @@ public class TestTask implements Serializable {
      */
     @XmlElementWrapper(name = "participants", required = true)
     @XmlElement(name = "participant")
+    @Singular
     private Set<TestParticipant> testParticipants;
 
     public TestTask() {